+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# <a href='https://github.com/mrjackwills/flightbox_vue/releases/tag/v0.1.4'>v0.1.4</a>
+### 2023-06-07
+
+### Chores
++ dependencies updated, [ec0f30b4](https://github.com/mrjackwills/flightbox_vue/commit/ec0f30b40d1dcb82cf1a3739c5593342a2068df5), [766cd243](https://github.com/mrjackwills/flightbox_vue/commit/766cd243b1be1b6033f88e904e74b4277c66afd2)
+
+### Features
++ use vuetify platform for v-if thumbnail tooltip, [cb540217](https://github.com/mrjackwills/flightbox_vue/commit/cb5402173370e1bf1eb2f99dad695bb5f8caae94)
++ use unplugin-vue-components to autoimport, [d7aed978](https://github.com/mrjackwills/flightbox_vue/commit/d7aed978204d6dbc632c1ea6084552ac713badf5)
+
 # <a href='https://github.com/mrjackwills/flightbox_vue/releases/tag/v0.1.3'>v0.1.3</a>
 ### 2023-05-11